refactor(rules): complete removeSpells rule metadata and promote spell damage-type helpers
- Add static group/description and documentUuid-widget uuids field to
  RemoveSpellsRule so it renders first-class in the Rules Builder, matching
  its grantSpells/restrictSpellSchools siblings; add en.json keys and a
  co-located removeSpells.test.ts
- Promote SCHOOL_TO_DAMAGE_TYPE/schoolToDamageType/retagEffectsDamageType out
  of view/dialogs/spellGrantUtils into src/utils/spell/spellDamageType so the
  character document no longer imports from the view layer; move their tests
- Remove orphaned NIMBLE.spellGrants.convertedSpellLabel i18n key
- Tidy collectSpellRestrictions loop variable naming
