+++ src/documents/actor/character.ts
@@ -17,8 +17,8 @@ import findMissingLevelSelections, {
 } from '#utils/findMissingLevelSelections.ts';
 import { buildClassFeatureIndex } from '#utils/getClassFeatures.ts';
 import { getHighestSpellTier } from '#utils/spell/getHighestSpellTier.ts';
+import { retagEffectsDamageType, SCHOOL_TO_DAMAGE_TYPE } from '#utils/spell/spellDamageType.ts';
 import CharacterMetaConfigDialog from '#view/dialogs/CharacterMetaConfigDialog.svelte';
-import { retagEffectsDamageType, SCHOOL_TO_DAMAGE_TYPE } from '#view/dialogs/spellGrantUtils.ts';
 import getDeterministicBonus from '../../dice/getDeterministicBonus.ts';
 import { NimbleRoll } from '../../dice/NimbleRoll.js';
 import { HitDiceManager, incrementDieSize } from '../../managers/HitDiceManager.js';
