+++ src/view/dialogs/spellGrantUtils.ts
@@ -12,25 +12,6 @@ export interface SpellRemovalEntry {
 }
 
 /**
- * Maps a spell school to the damage type its spells deal. Used when retagging a
- * "keep and convert" exception spell so its elemental damage matches the new
- * school. Note the mapping is not identity: the Ice school deals `cold` damage.
- */
-export const SCHOOL_TO_DAMAGE_TYPE: Record<string, string> = {
-	fire: 'fire',
-	ice: 'cold',
-	lightning: 'lightning',
-};
-
-/** The elemental damage types that a conversion may rewrite. */
-const ELEMENTAL_DAMAGE_TYPES = new Set(Object.values(SCHOOL_TO_DAMAGE_TYPE));
-
-/** Resolves the damage type for a school, or null when the school is non-elemental. */
-export function schoolToDamageType(school: string): string | null {
-	return SCHOOL_TO_DAMAGE_TYPE[school] ?? null;
-}
-
-/**
  * Active spell-school restriction collected from restrictSpellSchools rules.
  * `allowedSchools` persists across levels (derived from all active rules), while
  * the exception fields describe the one-time "keep and convert" choice.
@@ -56,9 +37,10 @@ export function collectSpellRestrictions(rulesArrays: RulesArray[]): SpellRestri
 		for (const rule of rules) {
 			if (rule.type !== 'restrictSpellSchools') continue;
 			active = true;
-			for (const s of (rule.allowedSchools as string[] | undefined) ?? []) allowedSchools.add(s);
-			for (const s of (rule.exceptionFromSchools as string[] | undefined) ?? [])
-				exceptionFromSchools.add(s);
+			for (const school of (rule.allowedSchools as string[] | undefined) ?? [])
+				allowedSchools.add(school);
+			for (const school of (rule.exceptionFromSchools as string[] | undefined) ?? [])
+				exceptionFromSchools.add(school);
 			exceptionCount += (rule.exceptionCount as number | undefined) ?? 0;
 		}
 	}
@@ -69,43 +51,6 @@ export function collectSpellRestrictions(rulesArrays: RulesArray[]): SpellRestri
 		exceptionFromSchools: [...exceptionFromSchools],
 		exceptionCount,
 	};
-}
-
-/**
- * Deep-clones a spell's `activation.effects` array, remapping any elemental
- * damage type (fire/cold/lightning) to `toDamageType`. Non-elemental damage
- * (radiant, necrotic, physical, …) is left untouched. Recurses into the nested
- * `on.<outcome>` effect arrays produced by saving-throw and attack effects.
- */
-export function retagEffectsDamageType(effects: unknown, toDamageType: string): unknown[] {
-	if (!Array.isArray(effects)) return [];
-
-	const remapNode = (node: unknown): unknown => {
-		if (!node || typeof node !== 'object') return node;
-		const clone: Record<string, unknown> = { ...(node as Record<string, unknown>) };
-
-		if (
-			typeof clone.damageType === 'string' &&
-			ELEMENTAL_DAMAGE_TYPES.has(clone.damageType as string)
-		) {
-			clone.damageType = toDamageType;
-		}
-
-		if (clone.on && typeof clone.on === 'object') {
-			const on = clone.on as Record<string, unknown>;
-			const newOn: Record<string, unknown> = { ...on };
-			for (const [outcome, children] of Object.entries(on)) {
-				if (Array.isArray(children)) {
-					newOn[outcome] = children.map(remapNode);
-				}
-			}
-			clone.on = newOn;
-		}
-
-		return clone;
-	};
-
-	return effects.map(remapNode);
 }
 
 /** Result of processing grantSpells rules */
